migrated to use state stores

+++ src-ui/index.ts
@@ -27,17 +27,10 @@
 import { renderBaseElements, renderGetStarted, renderEditor } from "renderer";
 import { Database } from "database";
 import { createEvent } from "event";
-import { EditorStore, StatusStore } from "store";
-import type { Note } from "types";
+import { NoteStore, EditorStore, StatusStore } from "store";
 
 // top-level app state (keep as small as possible)
-// TODO: revisit notes and selectedNoteId state. Might be best to use a Proxy Store if it's used app-wide
-// this is already global state, so moving to a proxy wouldn't really change anything.
-// It would give me access to the state anywhere, so it'd be better.
-// that state is only ever set here, so it's easier to read
 let database: Database;
-let notes: Record<string, Note> = {};
-let selectedNoteId: null | string = null;
 
 /**
  * Keyboard events
@@ -61,27 +54,27 @@ window.addEventListener("refresh-client", async () => {
    * by default it's by created_at,
    * but can easily be extended to sort by any note data
    */
-  notes = await database.getAll();
-  if (!selectedNoteId) {
+  NoteStore.notes = await database.getAll();
+  if (!NoteStore.selectedNoteId) {
     /**
      * To start, not calling the db again to get the most recent note.
      * However, if slow downs become noticeable, this would be a place to optimize.
      */
-    const sortedNotes = Object.values(notes).sort((a, b) => {
+    const sortedNotes = Object.values(NoteStore.notes).sort((a, b) => {
       return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
     });
-    selectedNoteId = sortedNotes[0]?._id;
+    NoteStore.selectedNoteId = sortedNotes[0]?._id;
   }
   // reset global state
   EditorStore.isDirty = false;
 
-  await refreshClient({ notes, selectedNoteId });
+  await refreshClient();
 });
 
 window.addEventListener("create-note", async (event) => {
   const { title, content = "" } = (event as CustomEvent)?.detail?.note;
   const id = await database.put({ title, content });
-  selectedNoteId = id;
+  NoteStore.selectedNoteId = id;
   dispatchEvent(new Event("refresh-client"));
 });
 
@@ -90,21 +83,18 @@ window.addEventListener("save-note", async () => {
   dispatchEvent(new Event("refresh-client"));
 });
 
-window.addEventListener("delete-note", async (event) => {
-  const { id } = (event as CustomEvent)?.detail?.note;
-  // TODO:
-  // move the fetch to the deleteNote route
-  // so we don't have to know the full state here
-  const noteToDelete = notes[id];
+window.addEventListener("delete-note", async () => {
+  if (!NoteStore.selectedNoteId) throw new Error("No note selected to delete");
+  const noteToDelete = NoteStore.notes[NoteStore.selectedNoteId];
   await database.delete(noteToDelete);
-  selectedNoteId = null; // reset selected note as it was deleted
+  NoteStore.selectedNoteId = null; // reset selected note as it was deleted
   dispatchEvent(new Event("refresh-client"));
 });
 
 window.addEventListener("select-note", async (event) => {
   if (EditorStore.isDirty) await saveNote();
   const { id } = (event as CustomEvent)?.detail?.note;
-  selectedNoteId = id;
+  NoteStore.selectedNoteId = id;
   dispatchEvent(new Event("refresh-client"));
 });
 
@@ -122,49 +112,42 @@ window.addEventListener("DOMContentLoaded", async () => {
 /**
  * Main app lifecycle, refresh based on latest state
  */
-async function refreshClient({
-  notes,
-  selectedNoteId,
-}: {
-  notes: Record<string, Note>;
-  selectedNoteId: string;
-}): Promise<void> {
+async function refreshClient(): Promise<void> {
   /**
    * Update state for initial render
    */
-  if (selectedNoteId) {
-    StatusStore.lastSavedDate = notes[selectedNoteId].updatedAt;
+  if (NoteStore.selectedNoteId) {
+    StatusStore.lastSavedDate =
+      NoteStore.notes[NoteStore.selectedNoteId].updatedAt;
   }
   const { editorElement, editorTopMenuElement, editorFloatingMenuElement } =
-    renderBaseElements(notes);
+    renderBaseElements(NoteStore.notes);
 
   // set main element content based on note state
-  if (!Object.keys(notes).length) {
+  if (!Object.keys(NoteStore.notes).length || !NoteStore.selectedNoteId) {
     StatusStore.lastSavedDate = null;
     renderGetStarted(editorElement);
     return;
   }
 
-  // NOTE: so if the
-
   EditorStore.editor = await renderEditor({
-    selectedNoteId,
     editorElement: editorElement,
     topEditorMenu: editorTopMenuElement,
     floatingEditorMenu: editorFloatingMenuElement,
-    editorContent: notes[selectedNoteId]?.content,
+    editorContent: NoteStore.notes[NoteStore.selectedNoteId]?.content,
   });
 
+  // set which note in the note-list is active
   toggleActiveClass({
-    selector: `#${selectedNoteId}-note-select-container`,
+    selector: `#${NoteStore.selectedNoteId}-note-select-container`,
     type: "select-note",
   });
 }
 
 async function saveNote() {
-  if (!selectedNoteId || !EditorStore.editor)
+  if (!NoteStore.selectedNoteId || !EditorStore.editor)
     throw new Error("No note selected to save or no editor instance");
-  const note = notes[selectedNoteId];
+  const note = NoteStore.notes[NoteStore.selectedNoteId];
   const content = EditorStore.editor.getHTML();
   note.content = content;
   await database.put(note);
